requirements.txt
discord.py>=2.0.0
PyNaCl>=1.5.0
ffmpeg-python>=0.2.0
